requirements.txt
onnxruntime>=1.18
numpy>=1.24
Pillow>=10.0
